The General Architecture
This is the General Architecture of my Project. 
This explains how all the components mentioned in the read me work together and what is the purpose of each of the components.
This also dives a bit deeper and tells you what the workflow inside N8N looks like and how they connect to each other.